feature: move compute_broadcast_shape to inferrer
