+++ dipu/torch_dipu/csrc_dipu/aten/ops/OpUtils.hpp
@@ -255,9 +255,5 @@ inline bool is_scalar_on_cpu(const at::Tensor& t) {
   return t.unsafeGetTensorImpl()->is_wrapped_number();
 }
 
-// coumpute broadcast shape based on two inputs
-// for example: a = [2, 1, 3], b = [2, 1], the result shape would be [2, 2, 3]
-at::DimVector compute_broadcast_shape(c10::IntArrayRef a, c10::IntArrayRef b);
-
 }  // namespace native
 }  // namespace dipu
